remove DEFAULT_PANNER_NODE_CONFIG
