+++ services/conference/src/scripts/models/audio/ConnectedGroup.ts
@@ -1,12 +1,10 @@
 import {convertToAudioCoordinate, getRelativePose} from '@models/utils'
+import {stereoParametersStore} from '@stores/AudioParameters'
 import {LocalParticipant} from '@stores/participants/LocalParticipant'
 import {RemoteParticipant} from '@stores/participants/RemoteParticipant'
 import {JitsiTrack} from 'lib-jitsi-meet'
 import {autorun, IObservableValue, IReactionDisposer} from 'mobx'
 import {NodeGroup} from './NodeGroup'
-
-import {stereoParametersStore} from '@stores/AudioParameters'
-import {ConfigurableProp} from '@stores/AudioParameters/StereoParameters'
 
 export class ConnectedGroup {
   private readonly disposers: IReactionDisposer[] = []
@@ -27,12 +25,8 @@ export class ConnectedGroup {
       },
     ))
 
-    const observedPannerKeys: ConfigurableProp[] =
-      ['coneInnerAngle', 'coneOuterAngle', 'coneOuterGain', 'distanceModel', 'maxDistance', 'distanceModel', 'panningModel', 'refDistance', 'rolloffFactor']
     this.disposers.push(autorun(
-      () => observedPannerKeys.forEach((key) => {
-        (group.pannerNode[key] as any) = stereoParametersStore[key]
-      }),
+      () => group.updatePannerConfig(stereoParametersStore),
     ))
   }
