+++ services/conference/src/scripts/models/audio/StereoParameters.ts
@@ -1,0 +1,4 @@
+export type ConfigurableParams = Pick<PannerNode, ConfigurableProp>
+
+export type ConfigurableProp = 'coneInnerAngle' | 'coneOuterAngle' | 'coneOuterGain' | 'distanceModel' |
+                         'maxDistance' | 'distanceModel' | 'panningModel' | 'refDistance' | 'rolloffFactor'
